% Generated by roxygen2 (4.0.1): do not edit by hand
\name{difftest}
\alias{difftest}
\title{difftest}
\usage{
difftest(data, pseudotime, df = 3)
}
\arguments{
\item{data}{The raw single_cell data, which is a numeric matrix or data.frame. Rows represent genes/features and columns represent single cells.}

\item{pseudotime}{The pseudotime information. It is typically the first element of the return value of function \code{\link{TSPpseudotime}}.}

\item{df}{Numeric value specifying the degree of freedom used in the GAM model.}
}
\value{
Data frame containing pvalues and qvalues of testing differentially expression.
}
\description{
testing differentially expressed genes
}
\details{
This function tests whether a gene is significantly expressed given pseudotime ordering. Generalized additive model (GAM) with user-specified
degrees of freedoms is compared with a constant fit to get the p-values. The p-values are adjusted for multiple testing using fdr to gain qvalues.
}
\examples{
data(lpsdata)
procdata <- preprocess(lpsdata)
#Choose STAT2 gene expression as marker gene
STAT2expr <- log2(lpsdata["STAT2",]+1)
lpspseudotime <- TSPpseudotime(procdata, geneexpr = STAT2expr, dim = 2)
diffval <- difftest(procdata,lpspseudotime[[1]])
#Selected differentially expressed genes under qvlue cutoff of 0.05
row.names(diffval)[diffval$qval < 0.05]
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}
\seealso{
\code{\link{TSPpseudotime}} for examples
}

