% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedynamics.R
\name{genedynamics}
\alias{genedynamics}
\title{genedynamics}
\usage{
genedynamics(geneexpr, order, k = 3)
}
\arguments{
\item{geneexpr}{A named gene expression vector. Names are cell names.}

\item{order}{A character vector of pseudotime cell ordering. Must be within the column names of \code{data}.}

\item{k}{degree of gam model. Smaller k makes the curve smoother.}
}
\value{
ggplot2 object.
}
\description{
plot gene expression values of individual genes along pseudotime
}
\details{
This function plots the expression values of individual genes against given pseudotime
}
\examples{
data(lpsdata)
procdata <- preprocess(lpsdata)
lpsmclust <- exprmclust(procdata)
lpsorder <- TSCANorder(lpsmclust,orderonly=TRUE,flip=TRUE)
#Choose STAT1 gene expression to plot
STAT2expr <- log2(lpsdata["STAT2",]+1)
singlegeneplot(STAT2expr, lpsorder)
}
\author{
Zhicheng Ji <zhicheng.ji@duke.edu>
}
