% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guided_tscan.R
\name{guided_tscan}
\alias{guided_tscan}
\title{guided_tscan}
\usage{
guided_tscan(dr, clu, cluorder)
}
\arguments{
\item{dr}{A numeric matrix of reduced dimensions. Can be from PCA, t-sne, UMAP, diffusion map, and many others. Each row is a cell and each column is a reduced dimension. Row names should be the cell names.}

\item{clu}{A named vector of cell clustering. Names are the cell names. Need to correspond to dr.}

\item{cluorder}{A vector indicating the order of cell clusters to be linked. Can be a subset of all clusters. Values in cluorder must be included in clu.}
}
\value{
A character vector of cell pseudotime ordering.
}
\description{
Guided pseudotime construction using TSCAN
}
\details{
This function should be used when users have already performed PCA or other dimension reductions, and cell clustering. Users also need to know which cell clusters should be linked in what order, which depends either on prior biological knowledge or obtained by guided_MST in an unsupervised fashion. This function fits well with any standard analysis pipeline such as Seurat, and can even accompany other pseudotime analysis methods such as diffusion maps. The biggest advantage is its flexibility. Users can supply any dimension reduction result, any cell clustering, and any order cell clusters should be linked. But it's users' responsibility to make sure the constructed pseudotime is biologically meaningful and interpreted in the correct way.
}
\examples{
# see vignette
}
\author{
Zhicheng Ji <zhicheng.ji@duke.edu>
}
