% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orderscore}
\alias{orderscore}
\title{orderscore}
\usage{
orderscore(subpopulation, orders)
}
\arguments{
\item{subpopulation}{Data frame with two columns. First column: cell names. Second column: sub-population codes.}

\item{orders}{A list with various length containing pseudotime orderings. Each pseudotime ordering is typically the first element of the return value of function \code{\link{TSPpseudotime}}.}
}
\value{
a numeric vector of calculated POS.
}
\description{
Calculate pseudotemporal ordering scores for orders
}
\details{
This function calculates pseudotemporal ordering scores (POS) based on the sub-population information and order information given by users.
Exactly two sub-population indicating early and late time points should be given. The early sub-population will be coded as 0 while the
late sub-population will be coded as 1.
}
\examples{
library(HSMMSingleCell)
library(Biobase)
data(HSMM)
HSMMdata <- exprs(HSMM)
HSMMdata <- HSMMdata[,grep("T0|72",colnames(HSMMdata))]
procdata <- preprocess(HSMMdata)
subpopulation <- data.frame(cell = colnames(procdata), sub = ifelse(grepl("T0",colnames(procdata)),0,1), stringsAsFactors = FALSE)
MYOGexpr <- log2(HSMMdata["ENSG00000122180.4",]+1)
#Comparing ordering with or without marker gene information
order1 <- TSPpseudotime(procdata, geneexpr = MYOGexpr, dim = 2)
order2 <- TSPpseudotime(procdata, dim = 2)
orders <- list(order1[[1]],order2[[1]])
orderscore(subpopulation, orders)
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

