% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotpseudotime}
\alias{plotpseudotime}
\title{plotpseudotime}
\usage{
plotpseudotime(pseudotimedata, x = 1, y = 2, show_tree = T,
  show_cell_names = T, cell_name_size = 3, markerexpr = NULL)
}
\arguments{
\item{pseudotimedata}{The exact output from \code{\link{TSPpseudotime}}.}

\item{x}{The column of data after dimension reduction to be plotted on the horizontal axis.}

\item{y}{The column of data after dimension reduction to be plotted on the vertical axis.}

\item{show_tree}{Whether to show the links between cells connected in the minimum spanning tree.}

\item{show_cell_names}{Whether to draw the name of each cell in the plot.}

\item{cell_name_size}{The size of cell name labels if show_cell_names is TRUE.}

\item{markerexpr}{The gene expression used to define the size of nodes.}
}
\value{
A ggplot2 object.
}
\description{
Plot the TSP constructed pseudotime time
}
\details{
This function will plot the gene expression data after dimension reduction and link the data points with the constructed pseudotime path. It is written by plot_spanning_tree functioin in package monocle.
}
\examples{
library(HSMMSingleCell)
library(Biobase)
data(HSMM)
HSMMdata <- exprs(HSMM)
HSMMdata <- HSMMdata[,grep("T0|72",colnames(HSMMdata))]
procdata <- preprocess(HSMMdata)
#Choose MYOG gene expression as marker gene
MYOGexpr <- log2(HSMMdata["ENSG00000122180.4",]+1)
HSMMpseudotime <- TSPpseudotime(procdata, geneexpr = MYOGexpr, dim = 2)
plotpseudotime(HSMMpseudotime, markerexpr = MYOGexpr)
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}
\references{
Cole Trapnell and Davide Cacchiarelli et al (2014): The dynamics and regulators of cell fate decisions are revealed by pseudo-temporal ordering of single cells. Nature Biotechnology
}
\seealso{
\code{\link{TSPpseudotime}} for examples
}

