% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readexpr.R
\name{readexpr}
\alias{readexpr}
\title{readexpr}
\usage{
readexpr(file, sep = ",")
}
\arguments{
\item{file}{The path to the file. The file contains the gene expression matrix. Note that the first column must be gene names. The first row (header) must be the cell names.}

\item{sep}{how the file is separated. For csv file, it is "," separated. For txt file, it is usually "\t" or " " separated.}
}
\value{
Matrix of gene expression.
}
\description{
read in the expression data and convert into a matrix
}
\details{
This function reads in the expression data and convert into a matrix.
}
\examples{
\dontrun{
exprdata <- readexpr("expr.txt",sep="\\t")
}
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}
