% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEdges.R
\name{reportEdges}
\alias{reportEdges}
\alias{reportEdges,ANY-method}
\alias{reportEdges,SummarizedExperiment-method}
\alias{reportEdges,SingleCellExperiment-method}
\title{Report edge coordinates}
\usage{
reportEdges(x, ...)

\S4method{reportEdges}{ANY}(x, mst, clusters, combined = TRUE)

\S4method{reportEdges}{SummarizedExperiment}(x, ..., assay.type = "logcounts")

\S4method{reportEdges}{SingleCellExperiment}(x, ..., use.dimred = NULL)
}
\arguments{
\item{x}{A numeric matrix of coordinates where each row represents a cell/sample and each column represents a dimension
(usually a PC or another low-dimensional embedding, but features or genes can also be used).

Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} object
containing such a matrix in its \code{\link{assays}}, as specified by \code{assay.type}.
This will be transposed prior to use.

Alternatively, for \linkS4class{SingleCellExperiment}s, this matrix may be extracted from its \code{\link{reducedDims}},
based on the \code{use.dimred} specification.
In this case, no transposition is performed.

Alternatively, if \code{clusters=NULL}, a numeric matrix of coordinates for cluster centroids,
where each row represents a cluster and each column represents a dimension 
Each row should be named with the cluster name.
This mode can also be used with assays/matrices extracted from SummarizedExperiments and SingleCellExperiments.}

\item{...}{For the generic, further arguments to pass to the specific methods.

For the SummarizedExperiment method, further arguments to pass to the ANY method.

For the SingleCellExperiment method, further arguments to pass to the SummarizedExperiment method
(if \code{use.dimred} is specified) or the ANY method (otherwise).}

\item{mst}{A \link{graph} object containing a MST, typically the output of \code{createClusterMST(x, clusters)}.}

\item{clusters}{A factor-like object of the same length as \code{nrow(x)},
specifying the cluster identity for each cell in \code{x}.
If \code{NULL}, \code{x} is assumed to already contain coordinates for the cluster centroids.}

\item{combined}{Logical scalar indicating whether a single data.frame of edge coordinates should be returned.}

\item{assay.type}{An integer or string specifying the assay to use from a SummarizedExperiment \code{x}.}

\item{use.dimred}{An integer or string specifying the reduced dimensions to use from a SingleCellExperiment \code{x}.}
}
\value{
A data.frame containing the start and end coordinates of segments representing all the edges in \code{mst}.
If \code{combined=FALSE}, a list of two data.frames is returned where corresponding rows represent the start and end coordinates of the same edge.
}
\description{
Provides the coordinates of the start and end of every edge.
This is mostly useful for plotting purposes in \code{\link{segments}} or the equivalent \pkg{ggplot2} functionality.
}
\details{
\code{x} may also be \code{NULL}, in which case the center coordinates are obtained 
from the \code{coordinates} vertex attribute of \code{mst}.
}
\examples{
# Re-using the example from ?createClusterMST.
example(createClusterMST, ask=FALSE) 

# Plotting the MST on top of existing visualizations:
edges <- reportEdges(x=NULL, mst, combined=FALSE)
plot(cells[,1], cells[,2], col=clusters)
segments(edges$start$dim1, edges$start$dim2, edges$end$dim1, 
     edges$end$dim2, lwd=5)

# Use with coordinates other than those used to make the MST:
shifted.cells <- cells + 10

shift.edges <- reportEdges(shifted.cells, mst, 
    clusters=clusters, combined=FALSE)
plot(shifted.cells[,1], shifted.cells[,2], col=clusters)
segments(shift.edges$start$dim1, shift.edges$start$dim2, 
    shift.edges$end$dim1, shift.edges$end$dim2, lwd=5)

# Also works for ggplot2:
df <- data.frame(shifted.cells, cluster=factor(clusters))
shift.edges2 <- reportEdges(shifted.cells, mst, clusters=clusters)

library(ggplot2)
ggplot(df) +
   geom_point(aes(x=X1, y=X2, color=cluster)) + 
   geom_line(data=shift.edges2, mapping=aes(x=dim1, y=dim2, group=edge))

}
\references{
Ji Z and Ji H (2016).
TSCAN: Pseudo-time reconstruction and evaluation in single-cell RNA-seq analysis.
\emph{Nucleic Acids Res.} 44, e117
}
\author{
Aaron Lun
}
