% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmean.R
\name{rowmean}
\alias{rowmean}
\title{Compute column means of a matrix based on a grouping variable}
\usage{
rowmean(x, group, ...)
}
\arguments{
\item{x}{A numeric matrix or matrix-like object.}

\item{group}{A vector or factor specifying the group assignment for each row of \code{x}.}

\item{...}{Further arguments to pass to \code{\link{rowsum}}, e.g., \code{reorder}.}
}
\value{
A numeric matrix with one row per level of \code{group},
where the value for each column contains the average value across the subset of rows corresponding that level.
}
\description{
A utility that is equivalent to \code{\link{rowsum}}, but computes the mean instead of the sum of each subset of rows.
}
\examples{
x <- matrix(runif(100), ncol = 5)
group <- sample(1:8, 20, TRUE)
(xmean <- rowmean(x, group))

}
\author{
Aaron Lun
}
