% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{SCAtools}
\alias{SCAtools}
\alias{SCAtools-package}
\title{SCAtools: Single-Cell Analysis Tools}
\description{
This package provides essential tools used in analyzing data from single-cell experiments
}
\details{
SCAtools enables users to easily construct and tune pseudotemporal cell ordering as well as analyzing differentially expressed genes.
SCAtools comes with a user-friendly GUI written in shiny. More functions will come in the future.
}

