% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TSCAN.R
\docType{package}
\name{TSCAN}
\alias{TSCAN}
\alias{TSCAN-package}
\title{TSCAN: Tools for Single-Cell ANalysis}
\description{
This package provides essential tools used in analyzing data from single-cell experiments
}
\details{
TSCAN enables users to easily construct and tune pseudotemporal cell ordering as well as analyzing differentially expressed genes.
TSCAN comes with a user-friendly GUI written in shiny. More functions will come in the future.
}

