% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TSCANui}
\alias{TSCANui}
\title{TSCANui}
\usage{
TSCANui()
}
\description{
Launch the TSCAN user interface in local machine
}
\details{
This function will automatically launch the TSCAN user interface in a web browser.
The user interface provides many powerful functions which is not available by command line programming.
It also provides a much easier and more convenient way to quickly explore single cell data and construct pseudotime analysis.
The user interface can also be accessed by http://zhiji.shinyapps.io/TSCAN. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally for faster running speed.
}
\examples{
\dontrun{
   TSCANui()
}
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

