% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TSPpseudotime}
\alias{TSPpseudotime}
\title{TSPpseudotime}
\usage{
TSPpseudotime(data, dim = "auto", statenum = 3, scale = TRUE,
  startpoint = NULL, flip = FALSE, geneexpr = NULL,
  exprtrend = "increasing", maxtime = 100, kmeansiter = 10)
}
\arguments{
\item{data}{The raw single_cell data, which is a numeric matrix or data.frame. Rows represent genes/features and columns represent single cells.}

\item{dim}{Either "auto" or a numeric value specifying the reduced dimenionality of PCA. If "auto" the optimal dimension will be chosen automatically.}

\item{statenum}{Numeric value specifiying number of cell states in K-means clustering.}

\item{scale}{Whether to scale gene expressions across all cells to have zero mean and unit variance. Normally this argument should be TRUE.}

\item{startpoint}{Manually specify the starting point of TSP ordering. Should be one of the column names of data. Omitted when geneexpr is not null.}

\item{flip}{Logical value specifying whether to flip the ordering.}

\item{geneexpr}{Gene expression values used to determine optimal starting point. Gene expression values should change monotonically in the true biological process. Names should agree exactly with column names of data (can be of different order).}

\item{exprtrend}{Trend of gene expression values. Either increasing or decreasing.}

\item{maxtime}{Numeric value to specify the pseudotime of the ending point.}

\item{kmeansiter}{Number of iterations of K-means clustering. The function will automatically pick an optimal clustering.}
}
\value{
a list containing
\itemize{
   \item pseudotime Data frame containing the constructed pseudotime information. First column: cell name. Second column: cell states. Third column: Pseudotime.
   \item reduceres Matrix storing the gene expression data after dimension reduction using PCA.
}
}
\description{
Construct pseudotime using Travelling Salesman Problem (TSP) algorithm
}
\details{
This function first uses principal component analysis (PCA) to reduce dimensionality of original data.
If not specified by the user, the optimal dimension will be automatically selected by fitting
a set of continuous piecewise regression to the standard deviations of first 20 principal components and choose
the one with the smallest residual sum of squares. The distance matrix between cells are calculated based on the reduced data.
Then the function uses nearest insertion algorithm to construct a TSP path as a suboptimal solution. Because TSP path does not
have a definite starting/ending point, users can specify a starting point, otherwise a random cell will be chosen as the starting point.
Users can also use the expression value of a gene to determine the optimal starting point. The gene expression value must change monotonically over the true biological process.
K-means clustering will be used to determine the different stages of cell during the biological process.
}
\examples{
data(lpsdata)
procdata <- preprocess(lpsdata)
#Choose STAT2 gene expression as marker gene
STAT2expr <- log2(lpsdata["STAT2",]+1)
TSPpseudotime(procdata, geneexpr = STAT2expr, dim = 2)
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}
\references{
Rosenkrantz, D. J., Stearns, R. E., & Lewis, II, P. M. (1977). An analysis of several heuristics for the traveling salesman problem. SIAM journal on computing, 6(3), 563-581.
}

