% Generated by roxygen2 (4.0.1): do not edit by hand
\name{orderscore}
\alias{orderscore}
\title{orderscore}
\usage{
orderscore(subpopulation, orders)
}
\arguments{
\item{subpopulation}{Data frame with two columns. First column: cell names. Second column: sub-population codes.}

\item{orders}{A list with various length containing pseudotime orderings. Each pseudotime ordering is typically the first element of the return value of function \code{\link{TSPpseudotime}}.}
}
\value{
a numeric vector of calculated POS.
}
\description{
Calculate pseudotemporal ordering scores for orders
}
\details{
This function calculates pseudotemporal ordering scores (POS) based on the sub-population information and order information given by users.
Exactly two sub-population indicating early and late time points should be given. The early sub-population will be coded as 0 while the
late sub-population will be coded as 1.
}
\examples{
data(lpsdata)
procdata <- preprocess(lpsdata)
subpopulation <- data.frame(cell = colnames(procdata), sub = ifelse(grepl("Unstimulated",colnames(procdata)),0,1), stringsAsFactors = FALSE)
#Choose STAT2 gene expression as marker gene
STAT2expr <- log2(lpsdata["STAT2",]+1)
#Comparing ordering with or without marker gene information
order1 <- TSPpseudotime(procdata, geneexpr = STAT2expr, dim = 2)
order2 <- TSPpseudotime(procdata, dim = 2)
orders <- list(order1[[1]],order2[[1]])
orderscore(subpopulation, orders)
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

