% Generated by roxygen2 (4.0.1): do not edit by hand
\name{singlegeneplot}
\alias{singlegeneplot}
\title{singlegeneplot}
\usage{
singlegeneplot(geneexpr, pseudotime, cell_size = 2)
}
\arguments{
\item{geneexpr}{The gene expression values. Names should agree with the pseudotime information.}

\item{pseudotime}{The pseudotime information. It is typically the first element of the return value of function \code{\link{TSPpseudotime}}.}

\item{cell_size}{Size of cells in the plot.}
}
\value{
ggplot2 object.
}
\description{
plot expression values of individual genes against pseudotime axis
}
\details{
This function plots the expression values of individual genes against given pseudotime
}
\examples{
data(lpsdata)
procdata <- preprocess(lpsdata)
#Choose STAT2 gene expression as marker gene
STAT2expr <- log2(lpsdata["STAT2",]+1)
lpspseudotime <- TSPpseudotime(procdata, geneexpr = STAT2expr)
#Choose STAT1 gene expression to plot
STAT1expr <- log2(lpsdata["STAT1",]+1)
singlegeneplot(STAT1expr, lpspseudotime[[1]])
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}
\seealso{
\code{\link{TSPpseudotime}} for examples
}

