% Generated by roxygen2 (4.0.2): do not edit by hand
\name{singlegeneplot}
\alias{singlegeneplot}
\title{singlegeneplot}
\usage{
singlegeneplot(geneexpr, pseudotime, cell_size = 2)
}
\arguments{
\item{geneexpr}{The gene expression values. Names should agree with the pseudotime information.}

\item{pseudotime}{The pseudotime information. It is typically the first element of the return value of function \code{\link{TSPpseudotime}}.}

\item{cell_size}{Size of cells in the plot.}
}
\value{
ggplot2 object.
}
\description{
plot expression values of individual genes against pseudotime axis
}
\details{
This function plots the expression values of individual genes against given pseudotime
}
\examples{
library(HSMMSingleCell)
library(Biobase)
data(HSMM)
HSMMdata <- exprs(HSMM)
HSMMdata <- HSMMdata[,grep("T0|72",colnames(HSMMdata))]
procdata <- preprocess(HSMMdata)
#Choose MYOG gene expression as marker gene
MYOGexpr <- log2(HSMMdata["ENSG00000122180.4",]+1)
HSMMpseudotime <- TSPpseudotime(procdata, geneexpr = MYOGexpr, dim = 2)
singlegeneplot(MYOGexpr, HSMMpseudotime[[1]])
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}
\seealso{
\code{\link{TSPpseudotime}} for examples
}

